removed casting to uint160 before casting to payable address

+++ contracts/zombiehelper.sol
@@ -14,7 +14,7 @@ contract ZombieHelper is ZombieFeeding {
   }
 
   function withdraw() external payable onlyOwner {
-    address payable _owner = payable(uint160(owner()));
+    address payable _owner = payable(owner());
     _owner.transfer(address(this).balance);
   }
